+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+* Expose user ID with getUserId.
+
 ## 1.7.0 (May 29, 2015)
 
 * Enable configuration of eventUploadThreshold, eventMaxCount,